server/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
python-dotenv==0.21.0
django-cors-headers==3.13.0
Pillow==9.3.0